exchange-sim/backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
psycopg[binary]>=3.1
channels>=4.0
channels-redis>=4.2
daphne>=4.0
requests>=2.31
python-dotenv>=1.0
gunicorn>=21.0
whitenoise[brotli]>=6.6.0
dj-database-url>=2.1
